+++ packages/isaacscript-common/src/features/extraConsoleCommands/listCommands.ts
@@ -45,6 +45,7 @@ import {
   LevelStage,
   PillColor,
   PillEffect,
+  PlayerForm,
   PlayerType,
   RoomType,
   SoundEffect,
@@ -70,7 +71,7 @@ import { addCharge } from "../../functions/charge";
 import { isValidCollectibleType } from "../../functions/collectibles";
 import { runDeepCopyTests } from "../../functions/deepCopyTests";
 import { getNPCs } from "../../functions/entitiesSpecific";
-import { getEnumValues } from "../../functions/enums";
+import { getEnumValues, getLastEnumValue } from "../../functions/enums";
 import { addFlag } from "../../functions/flag";
 import { spawnGridEntity } from "../../functions/gridEntities";
 import { getRoomGridIndexesForType } from "../../functions/levelGrid";
@@ -99,8 +100,13 @@ import { gridCoordinatesToWorldPosition } from "../../functions/roomGrid";
 import { changeRoom } from "../../functions/rooms";
 import { reloadRoom as reloadRoomFunction } from "../../functions/roomTransition";
 import { onSetSeed, restart, setUnseeded } from "../../functions/run";
+import { getSortedSetValues } from "../../functions/set";
 import { spawnCollectible as spawnCollectibleFunction } from "../../functions/spawnCollectible";
 import { setStage } from "../../functions/stage";
+import {
+  getCollectibleTypesForTransformation,
+  getTransformationName,
+} from "../../functions/transformations";
 import { getGoldenTrinketType } from "../../functions/trinkets";
 import {
   asCardType,
@@ -113,6 +119,7 @@ import { CHARACTER_NAME_TO_TYPE_MAP } from "../../maps/characterNameToTypeMap";
 import { COLLECTIBLE_NAME_TO_TYPE_MAP } from "../../maps/collectibleNameToTypeMap";
 import { PILL_NAME_TO_EFFECT_MAP } from "../../maps/pillNameToEffectMap";
 import { ROOM_NAME_TO_TYPE_MAP } from "../../maps/roomNameToTypeMap";
+import { TRANSFORMATION_NAME_TO_PLAYER_FORM_MAP } from "../../maps/transformationNameToPlayerForm";
 import { TRINKET_NAME_TO_TYPE_MAP } from "../../maps/trinketNameToTypeMap";
 import { getLastCardType, getLastPillEffect } from "../firstLast";
 import {
@@ -1000,6 +1007,11 @@ export function playSound(params: string): void {
   sound(params);
 }
 
+/** Alias for the "transformation" command. */
+export function playerForm(params: string): void {
+  transformation(params);
+}
+
 /** Sets the player's pocket item to the specified collectible type. */
 export function pocket(params: string): void {
   if (params === "") {
@@ -1454,6 +1466,60 @@ export function tears(): void {
 /** Alias for the "runTests" command. */
 export function tests(): void {
   runTests();
+}
+
+/**
+ * Gives the specified transformation. Accepts either the transformation number or the partial name
+ * of the transformation.
+ *
+ * For example:
+ * - transformation 1 - Gives the Beelzebub transformation.
+ * - transformation gup - Gives the Guppy transformation.
+ */
+export function transformation(params: string): void {
+  if (params === "") {
+    printConsole("You must specify a transformation name or number.");
+    return;
+  }
+
+  let targetPlayerForm: PlayerForm;
+  const num = tonumber(params) as PlayerForm | undefined;
+  if (num === undefined) {
+    const match = getMapPartialMatch(
+      params,
+      TRANSFORMATION_NAME_TO_PLAYER_FORM_MAP,
+    );
+    if (match === undefined) {
+      printConsole(`Unknown transformation: ${params}`);
+      return;
+    }
+
+    targetPlayerForm = match[1];
+  } else {
+    const lastPlayerForm = getLastEnumValue(PlayerForm);
+    if (num < PlayerForm.GUPPY || num > lastPlayerForm) {
+      printConsole(`Invalid transformation number: ${num}`);
+      return;
+    }
+
+    targetPlayerForm = num;
+  }
+
+  const transformationName = getTransformationName(targetPlayerForm);
+  const player = Isaac.GetPlayer();
+  const collectibleTypesSet =
+    getCollectibleTypesForTransformation(targetPlayerForm);
+  const collectiblesTypes = getSortedSetValues(collectibleTypesSet);
+  for (let i = 0; i < 3; i++) {
+    const collectibleType = collectiblesTypes[i];
+    if (collectibleType !== undefined) {
+      player.AddCollectible(collectibleType);
+    }
+  }
+
+  printConsole(
+    `Gave transformation: ${transformationName} (${targetPlayerForm})`,
+  );
 }
 
 /** Creates a trapdoor next to the player. */